requirements.txt
asgiref==3.8.1
Django==4.2.16
django-admin-interface==0.29.1
django-colorfield==0.11.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==11.0.0
PyJWT==2.9.0
python-slugify==8.0.4
sqlparse==0.5.1
text-unidecode==1.3
typing_extensions==4.12.2
